Added recursive call for error handling.
Sometimes, the HTML Parser would not correctly wait for the JavaScript to load. In this case, we have roughly a 6,000 line raw html return, and none of the foods were actually loaded. In the case of no problems, there are around 15k lines of raw html.

The if-statement checks if there are any classed foods in the received raw html. If there isn't any, the raw html is malformed, and it calls itself until that is no longer the case.